Update to fuser 0.6.0

+++ src/base/utils.rs
@@ -75,18 +75,18 @@ pub fn check_access(
     file_mode: u16,
     uid: u32,
     gid: u32,
-    mut access_mask: u32,
+    mut access_mask: i32,
 ) -> bool {
     // F_OK tests for existence of file
-    if access_mask == libc::F_OK as u32 {
+    if access_mask == libc::F_OK {
         return true;
     }
-    let file_mode = u32::from(file_mode);
+    let file_mode = i32::from(file_mode);
 
     // root is allowed to read & write anything
     if uid == 0 {
         // root only allowed to exec if one of the X bits is set
-        access_mask &= libc::X_OK as u32;
+        access_mask &= libc::X_OK;
         access_mask -= access_mask & (file_mode >> 6);
         access_mask -= access_mask & (file_mode >> 3);
         access_mask -= access_mask & file_mode;
